Refactor main; comments
